pyproject.toml
[project]
name = "agent-traveler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.19.0",
    "icalendar>=6.3.2",
    "pytz>=2025.2",
    "simplekml>=1.3.6",
]
